v6.1.0.1 Replaces mimemagic dependency with marcel
Given the licensing controversy, Rails has removed the dependency on mimemagic in favor of new functionality directly in the marcel library used by active storage.

https://weblog.rubyonrails.org/2021/3/26/marcel-upgrade-releases/

Paperclip is a deprecated gem that depends on mimemagic.
Because of this, even though we upgraded rails, we still depend on mimemagic.
We've circumvented the drama somewhat by pinning our mimemagic dependency to a SHA.

This patch replaces the mimemagic usage with analogous marcel functionality.

I've modified the version to 6.1.0.1 even though I don't intend on ever publishing this.
This patch is to be used by Root Insurance's monorepo via a forked version of paperclip.

This will be sunset when we prioritize removing paperclip entirely.